+++ pkg/source/applovin/applovin_test.go
@@ -1,9 +1,16 @@
 package applovin
 
 import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"strings"
 	"testing"
 	"time"
 
+	httpclient "github.com/bruin-data/ingestr/pkg/http"
+	"github.com/bruin-data/ingestr/pkg/source"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -268,5 +275,63 @@ func TestCreateCustomReportTable(t *testing.T) {
 				assert.NotNil(t, table)
 			}
 		})
+	}
+}
+
+func TestAppLovinByteCap(t *testing.T) {
+	wide := strings.Repeat("x", 2048)
+	const nRows = 50
+
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		rows := make([]map[string]interface{}, 0, nRows)
+		for i := 0; i < nRows; i++ {
+			rows = append(rows, map[string]interface{}{
+				"day":  "2024-01-01",
+				"blob": wide,
+			})
+		}
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(map[string]interface{}{"results": rows})
+	}))
+	defer srv.Close()
+
+	day := time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
+
+	run := func(maxBytes int64) (int64, int64) {
+		s := &AppLovinSource{
+			apiKey: "dummy",
+			client: httpclient.New(httpclient.WithBaseURL(srv.URL)),
+		}
+		results, err := s.readTable(context.Background(), "report", []string{"day", "blob"}, ReportTypeAdvertiser, source.ReadOptions{
+			IntervalStart: &day,
+			IntervalEnd:   &day,
+			MaxBatchBytes: maxBytes,
+		})
+		if err != nil {
+			t.Fatal(err)
+		}
+		var batches, rows int64
+		for res := range results {
+			if res.Err != nil {
+				t.Fatal(res.Err)
+			}
+			batches++
+			rows += res.Batch.NumRows()
+			res.Batch.Release()
+		}
+		return batches, rows
+	}
+
+	offB, offR := run(0)
+	onB, onR := run(4096)
+
+	if offB != 1 {
+		t.Fatalf("cap-off batches=%d want 1", offB)
+	}
+	if onB <= 1 {
+		t.Fatalf("cap-on batches=%d want >1", onB)
+	}
+	if offR != onR || offR != nRows {
+		t.Fatalf("row mismatch off=%d on=%d want %d", offR, onR, nRows)
 	}
 }
